Refactor: Move RangerModel and RangerView to correct architectural layers
